+++ crates/hardforks/src/hardfork/mod.rs
@@ -43,64 +43,3 @@ impl Hash for dyn Hardfork + 'static {
         self.name().hash(state)
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use alloc::vec::Vec;
-    use core::str::FromStr;
-
-    #[test]
-    fn check_hardfork_from_str() {
-        let hardfork_str = [
-            "frOntier",
-            "homEstead",
-            "dao",
-            "tAngerIne",
-            "spurIousdrAgon",
-            "byzAntium",
-            "constantinople",
-            "petersburg",
-            "istanbul",
-            "muirglacier",
-            "bErlin",
-            "lonDon",
-            "arrowglacier",
-            "grayglacier",
-            "PARIS",
-            "ShAnGhAI",
-            "CaNcUn",
-            "PrAguE",
-        ];
-        let expected_hardforks = [
-            EthereumHardfork::Frontier,
-            EthereumHardfork::Homestead,
-            EthereumHardfork::Dao,
-            EthereumHardfork::Tangerine,
-            EthereumHardfork::SpuriousDragon,
-            EthereumHardfork::Byzantium,
-            EthereumHardfork::Constantinople,
-            EthereumHardfork::Petersburg,
-            EthereumHardfork::Istanbul,
-            EthereumHardfork::MuirGlacier,
-            EthereumHardfork::Berlin,
-            EthereumHardfork::London,
-            EthereumHardfork::ArrowGlacier,
-            EthereumHardfork::GrayGlacier,
-            EthereumHardfork::Paris,
-            EthereumHardfork::Shanghai,
-            EthereumHardfork::Cancun,
-            EthereumHardfork::Prague,
-        ];
-
-        let hardforks: Vec<EthereumHardfork> =
-            hardfork_str.iter().map(|h| EthereumHardfork::from_str(h).unwrap()).collect();
-
-        assert_eq!(hardforks, expected_hardforks);
-    }
-
-    #[test]
-    fn check_nonexistent_hardfork_from_str() {
-        assert!(EthereumHardfork::from_str("not a hardfork").is_err());
-    }
-}
